pyproject.toml
[tool.poetry]
name = "python-tutorial"
version = "0.1.0"
description = ""
authors = ["ericcbonet <ericcbonet@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
grpcio = "^1.51.1"
grpcio-tools = "^1.51.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
